HW8: Реализовал поле search с помощью MUI
