+++ src/components/header/header.styles.scss
@@ -6,41 +6,10 @@
     justify-content: space-between;
     align-items: center;
 
-
-    /* Поле search */
-
     &__search-wrapper {
         display: flex;
         align-items: center;
     }
-
-    .search-field {
-        background: url('@svg/search.svg') no-repeat;
-        background-position: 3px center;
-        border: none;
-        border-bottom: 2px solid vars.$white;
-        width: 30rem;
-        height: 4rem;
-        font-size: 1.4rem;
-        padding-left: 3.2rem;
-
-        &::placeholder {
-            color: vars.$primary-color;
-        }
-
-        &:focus {
-            outline: none;
-            border-bottom: 2px solid vars.$active-color;
-            background: url('@svg/search_active.svg') no-repeat;
-            background-position: 3px center;
-
-            &::placeholder {
-                color: transparent;
-            }
-        }
-    }
-
-    /* Иконки корзины и аккаунта пользователя */
 
     &__icon-wrapper {
         display: flex;
