+++ src/components/product/quantity/quantity.styles.scss
