+++ src/slices/favoriteSlice/types.ts
@@ -1,0 +1,3 @@
+export interface IFavoritesState {
+    favoriteIds: string[];
+}
