+++ src/slices/quantitySlice/types.ts
@@ -1,0 +1,3 @@
+export interface IQuantityState {
+    value: number;
+}
